Makes Decimal a field on Transfer

+++ src/apps/chifra/pkg/ledger/get_transfers.go
@@ -57,7 +57,7 @@ func convertToTransfers(statements []types.Statement) ([]types.Transfer, error) 
 			BlockNumber:      stmnt.BlockNumber,
 			TransactionIndex: stmnt.TransactionIndex,
 			LogIndex:         stmnt.LogIndex,
-			Decimals:         stmnt.Decimals,
+			Decimals:         uint64(stmnt.Decimals),
 		}
 		if !t.Amount.Equal(base.ZeroWei) {
 			transfers = append(transfers, t)
